lives system implemented
