+++ Assets/HealthManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b02cdb5c27fc14f4ab6d2eed8f960ba2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
